Rename "StandardColumn" to "ColumnWithLabels"
