+++ include/AllegroFlare/Elements/RollingCredits/SectionFactory.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 
+#include <AllegroFlare/Elements/RollingCredits/Sections/ColumnWithLabels.hpp>
 #include <AllegroFlare/Elements/RollingCredits/Sections/Header.hpp>
-#include <AllegroFlare/Elements/RollingCredits/Sections/StandardColumn.hpp>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -22,7 +22,7 @@ namespace AllegroFlare
             SectionFactory();
             ~SectionFactory();
 
-            AllegroFlare::Elements::RollingCredits::Sections::StandardColumn build_standard_column_section(std::vector<std::tuple<std::string, std::string>> sections={});
+            AllegroFlare::Elements::RollingCredits::Sections::ColumnWithLabels build_standard_column_section(std::vector<std::tuple<std::string, std::string>> sections={});
             AllegroFlare::Elements::RollingCredits::Sections::Header build_header_section(std::string text="[unset-text]");
          };
       }
